// resources/js/types/FormBillingProfile.d.ts
export interface FormBillingProfile {
    legal_name: string;
    tax_id: string;
    billing_email: string;
    phone?: string;
    address_line_1: string;
    address_line_2?: string;
    city: string;
    state: string;
    postal_code: string;
    country: string;
    invoice_prefix: string;
    next_invoice_number: number;
    invoice_footer_notes?: string;
    logo_path?: string;
}